refactor: remove legacy internal module, migrate imports to repo_index::indexer and fix clippy warnings
